src/search-adaptor/src/requirements.txt
urllib3==1.26.5
Flask==1.1.2
jsonschema==3.2.0

# The commons package requires requests>=2.22.0 and PyYAML>=5.3.1
requests==2.25.1
PyYAML==5.4.1

# The branch name of commons to be used during image build:
# Default is master branch specified in docker-compose.yml if not set.
# For local development:
# COMMONS_BRANCH=master pip install -r requirements.txt
git+https://github.com/hubmapconsortium/commons.git@${COMMONS_BRANCH}#egg=hubmap-commons
